SDL.Filesystem

Filesystem Paths

getBasePath :: IO Text

An absolute path to the application data directory.

The path is guaranteed to end with a path separator.

Throws SDLException on failure, or if the platform does not implement this functionality.

getPrefPath :: Text -> Text -> IO Text

A path to a unique per user and per application directory for the given organization and application name, intended for writing preferences and other personal files.

The path is guaranteed to end with a path separator.

You should assume the path returned by this function is the only safe place to write files to.

Throws SDLException on failure.