SDL.Init

Documentation

init :: Foldable f => f InitFlag -> IO ()

Initializes SDL and the given subsystems. Do not call any SDL functions prior to this one, unless otherwise documented that you may do so.

Throws SDLException if initialization fails.

initSubSystem :: Foldable f => f InitFlag -> IO ()

Initialize individual subsystems. SDL needs to be initializied prior to calls to this function.

Throws SDLException if initialization fails.

data InitFlag

Constructors

InitTimer

InitAudio

InitVideo

InitJoystick

InitHaptic

InitGameController

InitEvents

InitEverything

InitNoParachute