SDL.Input.Keyboard

Keyboard Modifiers

getModState :: IO KeyModifier

Get the current key modifier state for the keyboard.

data KeyModifier

Constructors

KeyModifier

Fields

keyModifierLeftShift :: Bool

keyModifierRightShift :: Bool

keyModifierLeftCtrl :: Bool

keyModifierRightCtrl :: Bool

keyModifierLeftAlt :: Bool

keyModifierRightAlt :: Bool

keyModifierLeftGUI :: Bool

keyModifierRightGUI :: Bool

keyModifierNumLock :: Bool

keyModifierCapsLock :: Bool

keyModifierAltGr :: Bool

Text Input

startTextInput :: Rect -> IO ()

Set the rectangle used to type text inputs and start accepting text input events.

stopTextInput :: IO ()

Stop receiving any text input events.

Screen Keyboard

hasScreenKeyboardSupport :: IO Bool

Check whether the platform has screen keyboard support.

isScreenKeyboardShown :: Window -> IO Bool

Check whether the screen keyboard is shown for the given window.

Scancodes

getScancodeName :: Scancode -> IO String

data Scancode

Constructors

ScancodeUnknown

ScancodeA

ScancodeB

ScancodeC

ScancodeD

ScancodeE

ScancodeF

ScancodeG

ScancodeH

ScancodeI

ScancodeJ

ScancodeK

ScancodeL

ScancodeM

ScancodeN

ScancodeO

ScancodeP

ScancodeQ

ScancodeR

ScancodeS

ScancodeT

ScancodeU

ScancodeV

ScancodeW

ScancodeX

ScancodeY

ScancodeZ

Scancode1

Scancode2

Scancode3

Scancode4

Scancode5

Scancode6

Scancode7

Scancode8

Scancode9

Scancode0

ScancodeReturn

ScancodeEscape

ScancodeBackspace

ScancodeTab

ScancodeSpace

ScancodeMinus

ScancodeEquals

ScancodeLeftBracket

ScancodeRightBracket

ScancodeBackslash

ScancodeNonUSHash

ScancodeSemicolon

ScancodeApostrophe

ScancodeGrave

ScancodeComma

ScancodePeriod

ScancodeSlash

ScancodeCapsLock

ScancodeF1

ScancodeF2

ScancodeF3

ScancodeF4

ScancodeF5

ScancodeF6

ScancodeF7

ScancodeF8

ScancodeF9

ScancodeF10

ScancodeF11

ScancodeF12

ScancodePrintScreen

ScancodeScrollLock

ScancodePause

ScancodeInsert

ScancodeHome

ScancodePageUp

ScancodeDelete

ScancodeEnd

ScancodePageDown

ScancodeRight

ScancodeLeft

ScancodeDown

ScancodeUp

ScancodeNumLockClear

ScancodeKPDivide

ScancodeKPMultiply

ScancodeKPMinus

ScancodeKPPlus

ScancodeKPEnter

ScancodeKP1

ScancodeKP2

ScancodeKP3

ScancodeKP4

ScancodeKP5

ScancodeKP6

ScancodeKP7

ScancodeKP8

ScancodeKP9

ScancodeKP0

ScancodeKPPeriod

ScancodeNonUSBackslash

ScancodeApplication

ScancodePower

ScancodeKPEquals

ScancodeF13

ScancodeF14

ScancodeF15

ScancodeF16

ScancodeF17

ScancodeF18

ScancodeF19

ScancodeF20

ScancodeF21

ScancodeF22

ScancodeF23

ScancodeF24

ScancodeExecute

ScancodeHelp

ScancodeMenu

ScancodeSelect

ScancodeStop

ScancodeAgain

ScancodeUndo

ScancodeCut

ScancodeCopy

ScancodePaste

ScancodeFind

ScancodeMute

ScancodeVolumeUp

ScancodeVolumeDown

ScancodeKPComma

ScancodeEqualsAs400

ScancodeInternational1

ScancodeInternational2

ScancodeInternational3

ScancodeInternational4

ScancodeInternational5

ScancodeInternational6

ScancodeInternational7

ScancodeInternational8

ScancodeInternational9

ScancodeLang1

ScancodeLang2

ScancodeLang3

ScancodeLang4

ScancodeLang5

ScancodeLang6

ScancodeLang7

ScancodeLang8

ScancodeLang9

ScancodeAltErase

ScancodeSysReq

ScancodeCancel

ScancodeClear

ScancodePrior

ScancodeReturn2

ScancodeSeparator

ScancodeOut

ScancodeOper

ScancodeClearAgain

ScancodeCrSel

ScancodeExSel

ScancodeKP00

ScancodeKP000

ScancodeThousandsSeparator

ScancodeDecimalSeparator

ScancodeCurrencyUnit

ScancodeCurrencySubunit

ScancodeLeftParen

ScancodeRightParen

ScancodeLeftBrace

ScancodeRightBrace

ScancodeKPTab

ScancodeKPBackspace

ScancodeKPA

ScancodeKPB

ScancodeKPC

ScancodeKPD

ScancodeKPE

ScancodeKPF

ScancodeKPXOR

ScancodeKPPower

ScancodeKPPercent

ScancodeKPLess

ScancodeKPGreater

ScancodeKPAmpersand

ScancodeKPDBLAmpersand

ScancodeKPVerticalBar

ScancodeKPDBLVerticalBar

ScancodeKPColon

ScancodeKPHash

ScancodeKPSpace

ScancodeKPAt

ScancodeKPExclam

ScancodeKPMemStore

ScancodeKPMemRecall

ScancodeKPMemClear

ScancodeKPMemAdd

ScancodeKPMemSubtract

ScancodeKPMemMultiply

ScancodeKPMemDivide

ScancodeKPPlusMinus

ScancodeKPClear

ScancodeKPClearEntry

ScancodeKPBinary

ScancodeKPOctal

ScancodeKPDecimal

ScancodeKPHexadecimal

ScancodeLCtrl

ScancodeLShift

ScancodeLAlt

ScancodeLGUI

ScancodeRCtrl

ScancodeRShift

ScancodeRAlt

ScancodeRGUI

ScancodeMode

ScancodeAudioNext

ScancodeAudioPrev

ScancodeAudioStop

ScancodeAudioPlay

ScancodeAudioMute

ScancodeMediaSelect

ScancodeWWW

ScancodeMail

ScancodeCalculator

ScancodeComputer

ScancodeACSearch

ScancodeACHome

ScancodeACBack

ScancodeACForward

ScancodeACStop

ScancodeACRefresh

ScancodeACBookmarks

ScancodeBrightnessDown

ScancodeBrightnessUp

ScancodeDisplaySwitch

ScancodeKBDIllumToggle

ScancodeKBDIllumDown

ScancodeKBDIllumUp

ScancodeEject

ScancodeSleep

ScancodeApp1

ScancodeApp2

ScancodeNum

Keycodes

data Keycode

Constructors

KeycodeUnknown

KeycodeReturn

KeycodeEscape

KeycodeBackspace

KeycodeTab

KeycodeSpace

KeycodeExclaim

KeycodeQuoteDbl

KeycodeHash

KeycodePercent

KeycodeDollar

KeycodeAmpersand

KeycodeQuote

KeycodeLeftParen

KeycodeRightParen

KeycodeAsterisk

KeycodePlus

KeycodeComma

KeycodeMinus

KeycodePeriod

KeycodeSlash

Keycode0

Keycode1

Keycode2

Keycode3

Keycode4

Keycode5

Keycode6

Keycode7

Keycode8

Keycode9

KeycodeColon

KeycodeSemicolon

KeycodeLess

KeycodeEquals

KeycodeGreater

KeycodeQuestion

KeycodeAt

KeycodeLeftBracket

KeycodeBackslash

KeycodeRightBracket

KeycodeCaret

KeycodeUnderscore

KeycodeBackquote

KeycodeA

KeycodeB

KeycodeC

KeycodeD

KeycodeE

KeycodeF

KeycodeG

KeycodeH

KeycodeI

KeycodeJ

KeycodeK

KeycodeL

KeycodeM

KeycodeN

KeycodeO

KeycodeP

KeycodeQ

KeycodeR

KeycodeS

KeycodeT

KeycodeU

KeycodeV

KeycodeW

KeycodeX

KeycodeY

KeycodeZ

KeycodeCapsLock

KeycodeF1

KeycodeF2

KeycodeF3

KeycodeF4

KeycodeF5

KeycodeF6

KeycodeF7

KeycodeF8

KeycodeF9

KeycodeF10

KeycodeF11

KeycodeF12

KeycodePrintScreen

KeycodeScrollLock

KeycodePause

KeycodeInsert

KeycodeHome

KeycodePageUp

KeycodeDelete

KeycodeEnd

KeycodePageDown

KeycodeRight

KeycodeLeft

KeycodeDown

KeycodeUp

KeycodeNumLockClear

KeycodeKPDivide

KeycodeKPMultiply

KeycodeKPMinus

KeycodeKPPlus

KeycodeKPEnter

KeycodeKP1

KeycodeKP2

KeycodeKP3

KeycodeKP4

KeycodeKP5

KeycodeKP6

KeycodeKP7

KeycodeKP8

KeycodeKP9

KeycodeKP0

KeycodeKPPeriod

KeycodeApplication

KeycodePower

KeycodeKPEquals

KeycodeF13

KeycodeF14

KeycodeF15

KeycodeF16

KeycodeF17

KeycodeF18

KeycodeF19

KeycodeF20

KeycodeF21

KeycodeF22

KeycodeF23

KeycodeF24

KeycodeExecute

KeycodeHelp

KeycodeMenu

KeycodeSelect

KeycodeStop

KeycodeAgain

KeycodeUndo

KeycodeCut

KeycodeCopy

KeycodePaste

KeycodeFind

KeycodeMute

KeycodeVolumeUp

KeycodeVolumeDown

KeycodeKPComma

KeycodeKPEqualsAS400

KeycodeAltErase

KeycodeSysReq

KeycodeCancel

KeycodeClear

KeycodePrior

KeycodeReturn2

KeycodeSeparator

KeycodeOut

KeycodeOper

KeycodeClearAgain

KeycodeCrSel

KeycodeExSel

KeycodeKP00

KeycodeKP000

KeycodeThousandsSeparator

KeycodeDecimalSeparator

KeycodeCurrencyUnit

KeycodeCurrencySubunit

KeycodeKPLeftParen

KeycodeKPRightParen

KeycodeKPLeftBrace

KeycodeKPRightBrace

KeycodeKPTab

KeycodeKPBackspace

KeycodeKPA

KeycodeKPB

KeycodeKPC

KeycodeKPD

KeycodeKPE

KeycodeKPF

KeycodeKPXor

KeycodeKPPower

KeycodeKPPercent

KeycodeKPLess

KeycodeKPGreater

KeycodeKPAmpersand

KeycodeKPDblAmpersand

KeycodeKPVecticalBar

KeycodeKPDblVerticalBar

KeycodeKPColon

KeycodeKPHash

KeycodeKPSpace

KeycodeKPAt

KeycodeKPExclam

KeycodeKPMemStore

KeycodeKPMemRecall

KeycodeKPMemClear

KeycodeKPMemAdd

KeycodeKPMemSubtract

KeycodeKPMemMultiply

KeycodeKPMemDivide

KeycodeKPPlusMinus

KeycodeKPClear

KeycodeKPClearEntry

KeycodeKPBinary

KeycodeKPOctal

KeycodeKPDecimal

KeycodeKPHexadecimal

KeycodeLCtrl

KeycodeLShift

KeycodeLAlt

KeycodeLGUI

KeycodeRCtrl

KeycodeRShift

KeycodeRAlt

KeycodeRGUI

KeycodeMode

KeycodeAudioNext

KeycodeAudioPrev

KeycodeAudioStop

KeycodeAudioPlay

KeycodeAudioMute

KeycodeMediaSelect

KeycodeWWW

KeycodeMail

KeycodeCalculator

KeycodeComputer

KeycodeACSearch

KeycodeACHome

KeycodeACBack

KeycodeACForward

KeycodeACStop

KeycodeACRefresh

KeycodeACBookmarks

KeycodeBrightnessDown

KeycodeBrightnessUp

KeycodeDisplaySwitch

KeycodeKbdIllumToggle

KeycodeKbdIllumDown

KeycodeKbdIllumUp

KeycodeEject

KeycodeSleep

Keysym

data Keysym

Constructors

Keysym

Fields

keysymScancode :: Scancode

keysymKeycode :: Keycode

keysymModifier :: KeyModifier