SDL.Input.Mouse

Relative Mouse Mode

setRelativeMouseMode :: Bool -> IO ()

Sets the current relative mouse mode.

When relative mouse mode is enabled, cursor is hidden and mouse position will not change. However, you will be delivered relative mouse position change events.

Throws SDLException on failure.

getRelativeMouseMode :: IO Bool

Check if relative mouse mode is enabled.

Mouse and Touch Input

data MouseButton

Constructors

ButtonLeft

ButtonMiddle

ButtonRight

ButtonX1

ButtonX2

ButtonExtra !Int
An unknown mouse button.

data MouseDevice

Identifies what kind of mouse-like device this is.

Constructors

Mouse !Int
An actual mouse. The number identifies which mouse.

Touch
Some sort of touch device.

data MouseMotion

Are buttons being pressed or released?

Constructors

MouseButtonUp

MouseButtonDown

Warping the Mouse

data WarpMouseOrigin

warpMouse :: WarpMouseOrigin -> V2 CInt -> IO ()

Cursor Visibility

setCursorVisible :: Bool -> IO ()

getCursorVisible :: IO Bool