SDL.Video

Window Management

data Window

createWindow :: Text -> WindowConfig -> IO Window

Create a window with the given title and configuration.

Throws SDLException on failure.

defaultWindow :: WindowConfig

Default configuration for windows. Use the record update syntax to override any of the defaults.

data WindowConfig

Constructors

WindowConfig

Fields

windowBorder :: Bool

Defaults to True.

windowHighDPI :: Bool

Defaults to False. Can not be changed after window creation.

windowInputGrabbed :: Bool

Defaults to False. Whether the mouse shall be confined to the window.

windowMode :: WindowMode

Defaults to Windowed.

windowOpenGL :: Bool

Defaults to False. Can not be changed after window creation.

windowPosition :: WindowPosition

Defaults to Wherever.

windowResizable :: Bool

Defaults to False. Whether the window can be resized by the user. It is still possible to programatically change the size with setWindowSize.

windowSize :: V2 CInt

Defaults to (800, 600).

data WindowMode

Constructors

Fullscreen
Real fullscreen with a video mode change

FullscreenDesktop
Fake fullscreen that takes the size of the desktop

Maximized

Minimized

Windowed

data WindowPosition

Constructors

Centered

Wherever
Let the window mananger decide where it's best to place the window.

Absolute (Point V2 CInt)

destroyWindow :: Window -> IO ()

Destroy the given window. The Window handler may not be used afterwards.

Window Actions

hideWindow :: Window -> IO ()

raiseWindow :: Window -> IO ()

Raise the window above other windows and set the input focus.

showWindow :: Window -> IO ()

Window Attributes

setWindowBordered :: Window -> Bool -> IO ()

Set whether the window should have a border or not.

setWindowBrightness :: Window -> Float -> IO ()

Set the window's brightness, where 0.0 is completely dark and 1.0 is normal brightness.

Throws SDLException if the hardware does not support gamma correction, or if the system has run out of memory.

setWindowGammaRamp :: Window -> Maybe (Vector Word16) -> Maybe (Vector Word16) -> Maybe (Vector Word16) -> IO ()

setWindowGrab :: Window -> Bool -> IO ()

Set whether the mouse shall be confined to the window.

setWindowMode :: Window -> WindowMode -> IO ()

Change between window modes.

Throws SDLException on failure.

setWindowPosition :: Window -> WindowPosition -> IO ()

Set the position of the window.

setWindowSize :: Window -> V2 CInt -> IO ()

Set the size of the window. Values beyond the maximum supported size are clamped.

setWindowTitle :: Window -> Text -> IO ()

Set the title of the window.

Clipboard Handling

getClipboardText :: IO Text

Get the text from the clipboard.

Throws SDLException on failure.

hasClipboardText :: IO Bool

Checks if the clipboard exists, and has some text in it.

setClipboardText :: Text -> IO ()

Replace the contents of the clipboard with the given text.

Throws SDLException on failure.

Drawing Primitives

createTextureFromSurface :: Renderer -> Surface -> IO Texture

loadBMP :: FilePath -> IO Surface

renderClear :: Renderer -> IO ()

renderCopy :: Renderer -> Texture -> Maybe (Rectangle CInt) -> Maybe (Rectangle CInt) -> IO ()

renderDrawLine :: Renderer -> Point V2 CInt -> Point V2 CInt -> IO ()

renderDrawLines :: Renderer -> Vector (Point V2 CInt) -> IO ()

renderDrawPoint :: Renderer -> Point V2 CInt -> IO ()

renderDrawPoints :: Renderer -> Vector (Point V2 CInt) -> IO ()

renderDrawRect :: Renderer -> Rectangle CInt -> IO ()

renderDrawRects :: Renderer -> Vector (Rectangle CInt) -> IO ()

renderFillRect :: Renderer -> Maybe (Rectangle CInt) -> IO ()

renderFillRects :: Renderer -> Vector (Rectangle CInt) -> IO ()

renderPresent :: Renderer -> IO ()

renderSetClipRect :: Renderer -> Maybe (Rectangle CInt) -> IO ()

renderSetLogicalSize :: Renderer -> V2 CInt -> IO ()

renderSetScale :: Renderer -> V2 CFloat -> IO ()

renderSetViewport :: Renderer -> Maybe (Rectangle CInt) -> IO ()

setRenderDrawBlendMode :: Renderer -> BlendMode -> IO ()

setRenderDrawColor :: Renderer -> V4 Word8 -> IO ()

data BlendMode

Constructors

BlendNone

BlendAlphaBlend

BlendAdditive

BlendMod

data Rectangle a

Constructors

Rectangle (Point V2 a) (V2 a)

data Surface

data Texture

Display

getDisplays :: IO [Display]

Throws SDLException on failure.

data Display

Constructors

Display

Fields

displayName :: String

displayBoundsPosition :: Point V2 CInt

Position of the desktop area represented by the display, with the primary display located at (0, 0).

displayBoundsSize :: V2 CInt

Size of the desktop area represented by the display.

displayModes :: [DisplayMode]

data DisplayMode

Constructors

DisplayMode

Fields

displayModeFormat :: PixelFormat

displayModeSize :: V2 CInt

displayModeRefreshRate :: CInt

Display's refresh rate in hertz, or 0 if unspecified.

data VideoDriver

Constructors

VideoDriver

Fields

videoDriverName :: String

data PixelFormat

Constructors

Unknown

Index1LSB

Index1MSB

Index4LSB

Index4MSB

Index8

RGB332

RGB444

RGB555

BGR555

ARGB4444

RGBA4444

ABGR4444

BGRA4444

ARGB1555

RGBA5551

ABGR1555

BGRA5551

RGB565

BGR565

RGB24

BGR24

RGB888

RGBX8888

BGR888

BGRX8888

ARGB8888

RGBA8888

ABGR8888

BGRA8888

ARGB2101010

YV12

IYUV

YUY2

UYVY

YVYU

OpenGL

data GLAttribute

Constructors

GLRedSize

GLGreenSize

GLBlueSize

GLAlphaSize

GLBufferSize

GLDoubleBuffer

GLDepthSize

GLStencilSize

GLAccumRedSize

GLAccumGreenSize

GLAccumBlueSize

GLAccumAlphaSize

GLStereo

GLMultiSampleBuffers

GLMultiSampleSamples

GLAcceleratedVisual

GLRetainedBacking

GLContextMajorVersion

GLContextMinorVersion

GLContextFlags

GLContextProfileMask

GLShareWithCurrentContext

GLFramebufferSRGBCapable

GLContextEGL

glSetAttribute :: GLAttribute -> CInt -> IO ()

glSwapWindow :: Window -> IO ()

Replace the contents of the front buffer with the back buffer's. The contents of the back buffer are undefined, clear them with glClear or equivalent before drawing to them again.

glGetCurrentContext :: IO GLContext

glCreateContext :: Window -> IO GLContext

Create a new OpenGL context and makes it the current context for the window.

Throws SDLException if the window wasn't configured with OpenGL support, or if context creation fails.

glGetProcAddress :: CString -> IO (Ptr ())

data SwapInterval

Constructors

ImmediateUpdates

SynchronizedUpdates

LateSwapTearing

glSetSwapInterval :: SwapInterval -> IO ()

glResetAttributes :: IO ()

Screen Savers

Screen savers should be disabled when the sudden enablement of the monitor's power saving features would be inconvenient for when the user hasn't provided any input for some period of time, such as during video playback.

Screen savers are disabled by default upon the initialization of the video subsystem.

disableScreenSaver :: IO ()

Disable screen savers.

enableScreenSaver :: IO ()

Enable screen savers.

isScreenSaverEnabled :: IO Bool

Check whether screen savers are enabled.

Message Box

showSimpleMessageBox :: Maybe Window -> MessageKind -> Text -> Text -> IO ()

Show a simple message box with the given title and a message. Consider writing your messages to stderr too.

Throws SDLException if there are no available video targets.

data MessageKind

Constructors

Error

Warning

Information