SDL.Video.OpenGL

OpenGL

data GLAttribute

Constructors

GLRedSize

GLGreenSize

GLBlueSize

GLAlphaSize

GLBufferSize

GLDoubleBuffer

GLDepthSize

GLStencilSize

GLAccumRedSize

GLAccumGreenSize

GLAccumBlueSize

GLAccumAlphaSize

GLStereo

GLMultiSampleBuffers

GLMultiSampleSamples

GLAcceleratedVisual

GLRetainedBacking

GLContextMajorVersion

GLContextMinorVersion

GLContextFlags

GLContextProfileMask

GLShareWithCurrentContext

GLFramebufferSRGBCapable

GLContextEGL

glSetAttribute :: GLAttribute -> CInt -> IO ()

glSwapWindow :: Window -> IO ()

Replace the contents of the front buffer with the back buffer's. The contents of the back buffer are undefined, clear them with glClear or equivalent before drawing to them again.

glGetCurrentContext :: IO GLContext

glCreateContext :: Window -> IO GLContext

Create a new OpenGL context and makes it the current context for the window.

Throws SDLException if the window wasn't configured with OpenGL support, or if context creation fails.

glGetProcAddress :: CString -> IO (Ptr ())

data SwapInterval

Constructors

ImmediateUpdates

SynchronizedUpdates

LateSwapTearing

glSetSwapInterval :: SwapInterval -> IO ()

glResetAttributes :: IO ()