SDL.Video.OpenGL

OpenGL

data GLContext

glCreateContext :: Window -> IO GLContext

Create a new OpenGL context and makes it the current context for the window.

Throws SDLException if the window wasn't configured with OpenGL support, or if context creation fails.

data Profile

Constructors

Core Mode CInt CInt

Compatibility Mode CInt CInt

ES Mode CInt CInt

data Mode

Constructors

Normal

Debug

glMakeCurrent :: Window -> GLContext -> IO ()

Throws SDLException on failure.

glDeleteContext :: GLContext -> IO ()

glSwapWindow :: Window -> IO ()

Replace the contents of the front buffer with the back buffer's. The contents of the back buffer are undefined, clear them with glClear or equivalent before drawing to them again.

data SwapInterval

Constructors

ImmediateUpdates

SynchronizedUpdates

LateSwapTearing

glSetSwapInterval :: SwapInterval -> IO ()

Function Loading

glGetProcAddress :: CString -> IO (Ptr ())